+++ mods/src/multiplayer/inventory.cpp
@@ -115,6 +115,13 @@ static void ClientSideNetworkHandler_handle_SendInventoryPacket_injection(Client
     }
 }
 
+static void ClientSideNetworkHandler_handle_ContainerSetContentPacket_Inventory_setItem_injection(Inventory *self, int slot, ItemInstance *item_instance) {
+    if (item_instance && !item_instance->isNull()) {
+        // Don't set null items
+        self->setItem(slot, item_instance);
+    }
+}
+
 // Init
 void _init_multiplayer_inventory() {
     // Fix Opening Containers
@@ -132,8 +139,11 @@ void _init_multiplayer_inventory() {
     }
 
     // Receive Inventory
-    if (feature_has("Clear Existing Inventory When Overwritten By Server", server_disabled)) {
+    if (feature_has("Correctly Handle Empty Slots When Receiving Inventory Data", server_disabled)) {
         overwrite_calls(ClientSideNetworkHandler_handle_ContainerSetContentPacket, ClientSideNetworkHandler_handle_ContainerSetContentPacket_injection_inventory);
+
+        // Don't allow for "EmptyItemInstance.name<"
+        overwrite_call((void *) 0x6d3e8, Inventory_setItem, ClientSideNetworkHandler_handle_ContainerSetContentPacket_Inventory_setItem_injection);
     }
 
     // Send Inventory